chore(deps): bump jsonwebtoken 9→10, progenitor 0.9→0.11, aws-lc-sys 0.31→0.38
Security fixes:
- jsonwebtoken 10.3.0: fixes 2 MEDIUM security alerts (Type Confusion
  leading to potential authorization bypass)
- aws-lc-sys 0.38.0: fixes 3 HIGH security alerts (PKCS7_verify
  Signature/Certificate Chain Validation Bypass, AES-CCM Timing
  Side-Channel)

Breaking changes addressed:
- jsonwebtoken 10 requires explicit crypto backend (aws_lc_rs)
- jsonwebtoken 10 requires Clone on decoded types (Attestation)
- progenitor 0.11 generates int32 for OpenAPI int32 fields
  (AmountDisplay.decimals: i64 → i32)

+++ crates/yttrium/src/pay/mod.rs
@@ -635,7 +635,7 @@ impl From<CollectDataFieldResult> for types::CollectDataFieldResult {
 pub struct AmountDisplay {
     pub asset_symbol: String,
     pub asset_name: String,
-    pub decimals: i64,
+    pub decimals: i32,
     pub icon_url: Option<String>,
     pub network_icon_url: Option<String>,
     pub network_name: Option<String>,
